requirements.txt
Django==5.1.5
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
PyJWT==2.10.1
requests==2.32.3
pytest==8.3.4

